added template support. added template file. updated todo.txt

+++ futil.py
@@ -1,3 +1,4 @@
+#/usr/bin/python3
 #futil v1.0
 
 import os
@@ -11,7 +12,6 @@ subparsers = parser.add_subparsers(help="functions", dest="command")
 parser.add_argument('-d','--directory', type=str, metavar='<directory>', help='specify a directory. Default is current working directory')
 parser.add_argument('-v', '--verbose', action='store_true', help='toggle verbose output')
 parser.add_argument('-V', '--version', action='version', version='%(prog)s 1.0', help='show program\'s version number and exit')
-parser.add_argument('-a', "--aa", action="store_true") #monolithic help
 
 rename_parser = subparsers.add_parser("rename", help="rename files", usage="futil.py [-d] rename filetype newname [-p]")
 rename_parser.add_argument("filetype", type=str)
@@ -23,33 +23,21 @@ create_parser.add_argument("filetype", type=str)
 create_parser.add_argument("name", type=str)
 create_parser.add_argument("amount", type=int, help='number of files to create')
 create_parser.add_argument("-p", "--placeholder", type=int, metavar="<int>", help='number of placeholder digits. default is 4, maximum is 10')
-create_parser.add_argument("-t", "--template", action="store_true", help="define a template file")
+create_parser.add_argument("-t", "--template", type=str, metavar="<template file>", help="define a template file")
 
 remove_parser = subparsers.add_parser("remove", help="remove files", usage="futil.py [-d] remove filetype")
 remove_parser.add_argument("filetype", type=str)
 
-if len(sys.argv) == 1:
+if len(sys.argv) == 1: #if no args, print help message
     parser.print_help(sys.stderr)
-    exit(0)
+    exit(0) #exit with code 0
 
-args = parser.parse_args()
+args = parser.parse_args() #parse arguments into ns args
 
-if args.aa:
-    parser.print_help(sys.stdout)
-    subparsers_actions = [
-        action for action in parser._actions
-        if isinstance(action, argparse._SubParsersAction)]
-
-    for subparsers_action in subparsers_actions:
-        for choice, subparser in subparsers_action.choices.items():
-            print('{}'.format(choice))
-            print(subparser.format_help())
-    exit(1)
-
-if args.filetype[0] == '.' and args.filetype is not None:
+if args.filetype[0] == '.' and args.filetype is not None: #parsing for filetype. if "." in filetype, remove.
     args.filetype = args.filetype[1:]
 
-if args.directory is not None:
+if args.directory is not None: #check if path exists
     if not os.path.exists(args.directory):
         parser.error(f"Path {args.directory} does not exist.")
     else:
@@ -58,9 +46,9 @@ else:
     args.directory = os.getcwd()
     os.chdir(os.getcwd())
 
-if args.command == 'rename' or args.command == 'create':
+if args.command == 'rename' or args.command == 'create': #range checking
     if args.placeholder is None:
-        args.placeholder = 4
+        args.placeholder = 4 #default is 4
     elif args.placeholder > 10:
         parser.error("Placeholder must be less than 10.")
         exit(2)
@@ -88,14 +76,39 @@ if args.command == 'rename':
         exit(1)
 
 if args.command == 'create':
-    for x in range(args.amount):
-        filenum = str(x).zfill(args.placeholder)
-        file = args.name + str(filenum) + '.' + args.filetype
-        f = open(file, "w")
+    if args.template: 
+        if not os.path.exists(args.template): #catch for template file not existing
+            parser.error("Template file does not exist.")
+            exit(1)
+        with open(args.template) as t:
+            for x in range(args.amount):
+                filenum = str(x).zfill(args.placeholder)
+                if args.placeholder == 0 and args.amount == 1: #catch for single file 
+                    file = args.name + '.' + args.filetype
+                else:
+                    file = args.name + str(filenum) + '.' + args.filetype
+                f = open(file, "w")
+                for line in t:
+                    f.write(line)
+                f.close()
+                if args.verbose:
+                    print(f"Creating {file} from template {t}...")
+            print("Done.")
+        t.close()
+    else:
+        for x in range(args.amount):
+            filenum = str(x).zfill(args.placeholder)
+            if args.placeholder == 0:
+                file = args.name + '.' + args.filetype
+            else:
+                file = args.name + str(filenum) + '.' + args.filetype
+            f = open(file, "w")
+            f.close()
+            if args.verbose:
+                print(f'Creating {file}...')
+            print("Done.")
+
         f.close()
-        if args.verbose:
-            print(f'Creating {file}...')
-    print("Done.")
 
 if args.command == 'remove':
     for file in files:
